@@ -24,25 +24,25 @@
         "lint": "eslint . --ext .ts"
     },
     "dependencies": {
-        "@angular/animations": "~9.1.4",
-        "@angular/cdk": "9.2.2",
-        "@angular/common": "~9.1.4",
-        "@angular/compiler": "~9.1.4",
-        "@angular/core": "~9.1.4",
-        "@angular/flex-layout": "9.0.0-beta.29",
-        "@angular/forms": "~9.1.4",
-        "@angular/material": "9.2.2",
-        "@angular/platform-browser": "~9.1.4",
-        "@angular/platform-browser-dynamic": "~9.1.4",
-        "@angular/router": "~9.1.4",
-        "@microsoft/signalr": "3.1.3",
-        "@ngrx/effects": "9.1.0",
-        "@ngrx/store": "9.1.0",
-        "@ngrx/store-devtools": "9.1.0",
-        "@popperjs/core": "^2.4.0",
+        "@angular/animations": "~10.0.1",
+        "@angular/cdk": "10.0.0",
+        "@angular/common": "~10.0.1",
+        "@angular/compiler": "~10.0.1",
+        "@angular/core": "~10.0.1",
+        "@angular/flex-layout": "10.0.0-beta.32",
+        "@angular/forms": "~10.0.1",
+        "@angular/material": "10.0.0",
+        "@angular/platform-browser": "~10.0.1",
+        "@angular/platform-browser-dynamic": "~10.0.1",
+        "@angular/router": "~10.0.1",
+        "@microsoft/signalr": "3.1.5",
+        "@ngrx/effects": "9.2.0",
+        "@ngrx/store": "9.2.0",
+        "@ngrx/store-devtools": "9.2.0",
+        "@popperjs/core": "^2.4.2",
         "@types/common-tags": "^1.8.0",
-        "angular-auth-oidc-client": "11.0.0",
-        "bootstrap": "4.4.1",
+        "angular-auth-oidc-client": "11.1.3",
+        "bootstrap": "4.5.0",
         "core-js": "^2.6.5",
         "ie-shim": "0.1.0",
         "msgpack5": "4.2.1",
@@ -51,56 +51,56 @@
         "zone.js": "~0.10.3"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^0.901.4",
-        "@angular/cli": "~9.1.4",
-        "@angular/compiler-cli": "~9.1.4",
-        "@angular/language-service": "~9.1.4",
-        "@ngtools/webpack": "^9.1.4",
-        "@types/jasmine": "~3.5.10",
+        "@angular-devkit/build-angular": "^0.1000.0",
+        "@angular/cli": "~10.0.0",
+        "@angular/compiler-cli": "~10.0.0",
+        "@angular/language-service": "~10.0.0",
+        "@ngtools/webpack": "^10.0.0",
+        "@types/jasmine": "~3.5.11",
         "@types/jasminewd2": "~2.0.8",
-        "@types/node": "~13.13.4",
+        "@types/node": "~14.0.14",
         "angular-router-loader": "0.8.5",
         "angular2-template-loader": "^0.6.2",
         "awesome-typescript-loader": "^5.2.1",
-        "clean-webpack-plugin": "^2.0.2",
-        "codelyzer": "~5.2.1",
-        "concurrently": "^5.1.0",
+        "clean-webpack-plugin": "2.0.2",
+        "codelyzer": "~5.2.2",
+        "concurrently": "^5.2.0",
         "copy-webpack-plugin": "^5.1.1",
         "css-loader": "^3.4.2",
-        "eslint": "^6.8.0",
         "file-loader": "^3.0.1",
         "html-webpack-plugin": "^3.2.0",
+        "eslint": "^7.3.1",
         "jasmine-core": "~3.5.0",
         "jasmine-spec-reporter": "~5.0.2",
-        "jquery": "^3.5.0",
+        "jquery": "^3.5.1",
         "json-loader": "^0.5.7",
-        "karma": "~5.0.4",
+        "karma": "~5.1.0",
         "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage-istanbul-reporter": "~2.1.1",
-        "karma-jasmine": "~3.1.1",
-        "karma-jasmine-html-reporter": "^1.5.3",
+        "karma-coverage-istanbul-reporter": "~3.0.3",
+        "karma-jasmine": "~3.3.1",
+        "karma-jasmine-html-reporter": "^1.5.4",
         "karma-sourcemap-loader": "^0.3.7",
         "karma-spec-reporter": "^0.0.32",
         "karma-webpack": "4.0.2",
-        "node-sass": "^4.14.0",
         "protractor": "~6.0.0",
         "raw-loader": "^1.0.0",
+        "node-sass": "^4.14.1",
         "sass-loader": "^8.0.2",
         "rimraf": "^3.0.2",
         "source-map-loader": "^0.2.4",
         "style-loader": "^1.1.3",
         "toposort": "2.0.2",
         "url-loader": "^1.1.2",
-        "ts-node": "~8.9.1",
-        "@typescript-eslint/eslint-plugin-tslint": "^2.30.0",
-        "@typescript-eslint/eslint-plugin": "^2.30.0",
-        "@typescript-eslint/parser": "^2.30.0",
+        "ts-node": "~8.10.2",
+        "@typescript-eslint/eslint-plugin-tslint": "^3.4.0",
+        "@typescript-eslint/eslint-plugin": "^3.4.0",
+        "@typescript-eslint/parser": "^3.4.0",
         "uglifyjs-webpack-plugin": "^2.2.0",
-        "typescript": "~3.8.3",
+        "typescript": "~3.9.5",
         "webpack": "^4.43.0",
-        "webpack-bundle-analyzer": "^3.7.0",
-        "webpack-cli": "3.3.11",
-        "webpack-dev-server": "^3.10.3",
+        "webpack-bundle-analyzer": "^3.8.0",
+        "webpack-cli": "3.3.12",
+        "webpack-dev-server": "^3.11.0",
         "webpack-filter-warnings-plugin": "^1.2.1"
     },
     "-vs-binding": {
