@@ -10,7 +10,7 @@ import { Configuration } from '../app.constants';
 })
 
 export class CoreModule {
-    static forRoot(): ModuleWithProviders {
+    static forRoot(): ModuleWithProviders<CoreModule> {
         return {
             ngModule: CoreModule,
             providers: [
